feat: refactor session list into a component and fix apiUtils
